requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
Pillow>=9.0.0
requests>=2.25.0